Refactor Cart class to use CartItem type
The changes made in this commit focus on refactoring the Cart class in the ecommerce module. Prior to this change, products in the cart were represented as generic objects. With this update, products are now represented by the stronger CartItem type which should lead to clearer and safer code. The `addProduct` and `removeProduct` methods have also been updated accordingly to accept CartItems instead of generic objects.