Change properties of ecommerce objects to public
The properties of the Product, ProductCategory, and CartItem classes under the ecommerce module have been changed from private to public. This allows for more external manipulation and interaction with these properties. Additionally, sample usage of these objects has been added to the 'demo.html' script, where a dummy product is created and added to the cart.